+++ project/src/app/SignIn/components/SignInFormContainer.tsx
@@ -1,20 +1,19 @@
 "use client";
 
 import { signIn } from "@/share/util/authUtil";
-import { Button, Grid, Paper } from "@mui/material";
 import { FormProvider, useForm } from "react-hook-form";
 import Cookies from "js-cookie";
-import { FormEmailTextField } from "@/share/components/form/FormEmailTextField";
-import { FormPasswordTextField } from "@/share/components/form/FormPasswordTextField";
 import { isAxiosError } from "axios";
 import { useRouter } from "next/navigation";
+import { SignInFormPresentation } from "./SignInFormPresentation";
+import { useCallback } from "react";
 
 export type SignInFormType = {
   email: string;
   password: string;
 };
 
-export const SignInForm = () => {
+export const SignInFormContainer = () => {
   const router = useRouter();
   const signInForm = useForm<SignInFormType>({
     defaultValues: { email: "", password: "" },
@@ -24,7 +23,7 @@ export const SignInForm = () => {
   /* 
     「サインイン」ボタン押下時
   */
-  const handleSignInOnClick = async () => {
+  const handleSignInOnClick = useCallback(async (_event: React.MouseEvent<HTMLButtonElement>) => {
     try {
       const res = await signIn(signInForm.getValues());
       if (res.status === 200) {
@@ -58,49 +57,12 @@ export const SignInForm = () => {
         alert("不明なエラー");
       }
     }
-  };
+  }, []);
 
   return (
     <>
       <FormProvider {...signInForm}>
-        <Paper elevation={0} sx={{ height: "100vh" }}>
-          <Grid container className="mt-10">
-            <Grid item className="w-96">
-              <FormEmailTextField
-                name="email"
-                label="メールアドレス"
-                helperText={signInForm.formState.errors.email?.message}
-                InputLabelProps={{ className: "font-mono" }}
-              />
-            </Grid>
-          </Grid>
-          <Grid container className="mt-10">
-            <Grid item className="w-96">
-              <FormPasswordTextField
-                name="password"
-                label="パスワード"
-                helperText={signInForm.formState.errors.password?.message}
-                InputLabelProps={{ className: "font-mono" }}
-              />
-            </Grid>
-          </Grid>
-          <Grid container className="justify-center mt-16">
-            <Grid item>
-              <Button
-                className="text-gray-500 border-gray-500 font-mono"
-                variant="text"
-                disabled={false}
-                onClick={handleSignInOnClick}
-                sx={{
-                  height: "60px",
-                  width: "130px",
-                }}
-              >
-                サインイン
-              </Button>
-            </Grid>
-          </Grid>
-        </Paper>
+        <SignInFormPresentation handleSignInOnClick={handleSignInOnClick} />
       </FormProvider>
     </>
   );
