+++ project/src/app/SignIn/page.tsx
@@ -1,6 +1,6 @@
-import { Header } from "@/share/components/Header";
+import { HeaderContainer } from "@/share/components/HeaderContainer";
 import { Grid, Typography } from "@mui/material";
-import { SignInForm } from "./components/SignInForm";
+import { SignInFormContainer } from "./components/SignInFormContainer";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { checkAuthenticated } from "@/share/util/authUtil";
@@ -25,7 +25,7 @@ const SignIn = async () => {
 
   return (
     <>
-      <Header />
+      <HeaderContainer />
       <Grid container className="justify-center mt-10">
         <Grid item>
           <Typography className="text-gray-500 font-mono" variant="h4">
@@ -35,7 +35,7 @@ const SignIn = async () => {
       </Grid>
       <Grid container className="justify-center mt-5">
         <Grid item>
-          <SignInForm />
+          <SignInFormContainer />
         </Grid>
       </Grid>
     </>
