+++ project/src/share/components/HeaderPresentation.tsx
@@ -13,53 +13,18 @@ import {
   Typography,
 } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
-import { useCallback, useState } from "react";
-import { signOut } from "../util/authUtil";
 import Cookies from "js-cookie";
-import { useRouter } from "next/navigation";
-import { isAxiosError } from "axios";
 
-export const Header = () => {
-  const [isDrawerOpend, setIsDrawerOpend] = useState(false);
+type HeaderPresentationProps = {
+  isDrawerOpend: boolean;
+  handleDrawerOpenClose: () => void;
+  handleSignOutOnClick: () => void;
+};
 
-  const router = useRouter();
-
-  // Drawerの開閉イベント
-  const handleDrawerOpenClose = useCallback(() => {
-    if (isDrawerOpend) {
-      setIsDrawerOpend(false);
-    } else {
-      setIsDrawerOpend(true);
-    }
-  }, [isDrawerOpend]);
-
-  /*
-    「ログアウト」ボタン押下時
-  */
-  const handleSignOutOnClick = async () => {
-    try {
-      const res = await signOut();
-      if (res.status === 200) {
-        // ブラウザのcookiesに保存しているトークンを削除
-        Cookies.remove("_access-token");
-        Cookies.remove("_client");
-        Cookies.remove("_uid");
-
-        // サインインページへリダイレクト
-        router.push("/");
-      }
-    } catch (e) {
-      if (isAxiosError(e)) {
-        if (e.response?.status === 404) {
-          alert(e.response.data.errors);
-        }
-      } else {
-        alert("不明なエラーです");
-      }
-    }
-  };
-
+export const HeaderPresentation = (props: HeaderPresentationProps) => {
   console.log("Headers cookies", Cookies.get());
+
+  const { isDrawerOpend, handleDrawerOpenClose, handleSignOutOnClick } = props;
 
   const drawerList = (
     <Box sx={{ width: "350px", paddingTop: "25px" }}>
